Move Adapters::HTTP to its own file
